refactor: move docs loader into enterprise
